requirements.txt
setuptools==70.2.0
langchain-groq==0.1.5
langchain-openai==0.1.14
langchain_community==0.2.5
tiktoken==0.7.0
langchainhub==0.1.20
langchain==0.2.5
langgraph==0.0.69
tavily-python==0.3.3
trafilatura==1.7.0
beautifulsoup4==4.12.3
fake-useragent==1.5.1
openai==1.35.9
python-dotenv==1.0.1
pandas==2.2.1
pytest==8.2.2
argparse==1.4.0
requests==2.32.3
bleu==0.3
jsonlines==4.0.0
jinja2==3.1.4